requirements/base.txt
Django==1.8.6
Pygments==2.0.2
blessings==1.6
bpython==0.14.2
curtsies==0.1.19
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-python3-ldap==0.9.7
djangorestframework==3.2.5
djangorestframework-gis==0.9.4
greenlet==0.4.9
ldap3==0.9.8.7.post1
mock==2.0.0
model-mommy==1.2.5.1
pbr==1.10.0
psycopg2==2.6.1
pyasn1==0.1.9
requests==2.7.0
six==1.9.0
sqlparse==0.1.16
